cấu hình thư mục
