+++ src/Components/Footer/Footer.jsx
@@ -1,6 +1,6 @@
 import { View, Text ,StyleSheet, TouchableOpacity} from 'react-native'
 import React from 'react'
-import Colors from './../../Utils/Colors';
+import Colors from '../../Utils/Colors';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 export default function footer() {
   return (
