+++ src/Components/Layout/Form.jsx
@@ -1,4 +1,4 @@
-import Colors from './../Utils/Colors'
+import Colors from '../../Utils/Colors'
 
 module.exports = {
     head1: {
