--- src/screens/TaskScreen/Task.jsx
@@ -1,14 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
-
-const Task = () => {
-  return (
-    <View>
-      <Text>Task</Text>
-    </View>
-  )
-}
-
-export default Task
-
-const styles = StyleSheet.create({})
